properly destroy notebook process
(cherry picked from commit b74697e27edb6595b59f0cb77b9cc5850f819b87)

+++ python/ipnb/src/org/jetbrains/plugins/ipnb/configuration/IpnbConnectionManager.java
@@ -334,7 +334,7 @@ public final class IpnbConnectionManager implements ProjectComponent {
       }
     }
     myKernels.clear();
-    if (myProcessHandler != null) {
+    if (myProcessHandler != null && !myProcessHandler.isProcessTerminated()) {
       myProcessHandler.killProcess();
     }
   }
